+++ lib/libpthread/arch/mips64/uthread_machdep.h
@@ -1,6 +1,6 @@
-/*	$OpenBSD: uthread_machdep.h,v 1.1 2004/08/11 17:41:34 pefo Exp $	*/
+/*	$OpenBSD: uthread_machdep.h,v 1.2 2004/09/09 16:59:21 pefo Exp $	*/
 /* David Leonard, <d@csee.uq.edu.au>. Public domain. */
 
 struct _machdep_state {
-	int	frame;
+	long	frame;
 };
